backend/requirements.txt
cloudinary==1.36.0
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
marshmallow-sqlalchemy==0.29.0
python-dotenv==1.0.0
SQLAlchemy==2.0.23

